requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.9.1
    # via openai
aiosignal==1.3.1
    # via aiohttp
astroid==2.15.8
    # via pylint
attrs==23.2.0
    # via aiohttp
blinker==1.7.0
    # via flask
cachetools==5.3.2
    # via google-auth
certifi==2023.11.17
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.7
    # via flask
dill==0.3.7
    # via pylint
flask==2.3.2
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
google-api-core[grpc]==2.15.0
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-auth==2.26.2
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-aiplatform==1.39.0
    # via -r requirements.in
google-cloud-bigquery==3.16.0
    # via google-cloud-aiplatform
google-cloud-core==2.4.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-resource-manager==1.11.0
    # via google-cloud-aiplatform
google-cloud-storage==2.9.0
    # via
    #   -r requirements.in
    #   google-cloud-aiplatform
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.7.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.62.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.13.0
    # via google-cloud-resource-manager
grpcio==1.60.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.60.0
    # via google-api-core
idna==3.6
    # via
    #   requests
    #   yarl
importlib-metadata==7.0.1
    # via yapf
isort==5.13.2
    # via pylint
itsdangerous==2.1.2
    # via flask
jinja2==3.1.3
    # via flask
lazy-object-proxy==1.10.0
    # via astroid
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
mccabe==0.7.0
    # via pylint
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
nodeenv==1.8.0
    # via pyright
numpy==1.26.3
    # via
    #   pandas
    #   shapely
openai==0.27.8
    # via -r requirements.in
packaging==23.2
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
pandas==2.1.1
    # via -r requirements.in
platformdirs==4.1.0
    # via
    #   pylint
    #   yapf
proto-plus==1.23.0
    # via
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
protobuf==4.25.2
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pylint==2.17.5
    # via -r requirements.in
pyright==1.1.345
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   google-cloud-bigquery
    #   pandas
pytz==2023.3.post1
    # via pandas
pyyaml==6.0
    # via -r requirements.in
regex==2023.12.25
    # via tiktoken
requests==2.28.1
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   openai
    #   tiktoken
rsa==4.9
    # via google-auth
shapely==2.0.2
    # via google-cloud-aiplatform
six==1.16.0
    # via python-dateutil
tiktoken==0.5.1
    # via -r requirements.in
tomli==2.0.1
    # via yapf
tomlkit==0.12.3
    # via pylint
tqdm==4.66.1
    # via openai
tzdata==2023.4
    # via pandas
urllib3==1.26.18
    # via requests
werkzeug==3.0.1
    # via flask
wrapt==1.16.0
    # via astroid
yapf==0.40.1
    # via -r requirements.in
yarl==1.9.4
    # via aiohttp
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
